+++ app/src/main/java/woowacourse/payments/ui/addcard/AddCardActivity.kt
@@ -1,0 +1,16 @@
+package woowacourse.payments.ui.addcard
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+
+class AddCardActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            GenerateCardView()
+        }
+    }
+}
